backend/requirements.txt
django>=4.0
djangorestframework>=3.12
djangorestframework-simplejwt>=4.7
psycopg2-binary>=2.9